third_party/blink/web_tests/external/wpt/dom/parts/basic-dom-part-objects.tentative.html

<!DOCTYPE html>
<title>DOM Parts: Basic object structure</title>
<meta name="author" href="mailto:masonf@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div id=target style="display:none">
  Imperative test element
  <span></span> A
  <span></span> B
  <span></span> C
</div>

<div style="display:none">
  Declarative syntax
  <h1 id="name">First<?child-node-part name?>Middle<?/child-node-part?>Last</h1>
  Email: <?node-part email-link?><a id="link"></a>
</div>

<script>
const target = document.getElementById('target');
assert_true(target.children.length >= 3);

test(() => {
  const root = document.getPartRoot();
  assert_true(root instanceof DocumentPartRoot,'document.getPartRoot() should return the DocumentPartRoot');
  const parts = root.getParts();
  assert_true(!!parts,'getParts() should return something');

  const nodePart = new NodePart(root,target);
  assert_true(!!nodePart);
  assert_equals(nodePart.node,target);
  assert_equals(nodePart.root,root);
  const childNodePart = new ChildNodePart(root,target.children[0], target.children[1]);
  assert_true(!!childNodePart);
  assert_equals(childNodePart.root,root);
  assert_equals(childNodePart.previousSibling,target.children[0]);
  assert_equals(childNodePart.nextSibling,target.children[1]);
  assert_equals(childNodePart.getParts().length,0,'childNodePart.getParts() should return an empty array');
}, 'Basic imperative DOM Parts object construction');
</script>
